Map Subset Cost To Float.

+++ src/threonine/map/FolderUsage.java
@@ -4,7 +4,7 @@ public class FolderUsage {
     //======================================================================
     long projectid = 0;
     long folderid = 0;
-    int costperuse = 0;
+    float costperuse = 0;
     String publicname = null;
     String sharepass = null;
     //======================================================================
@@ -12,11 +12,11 @@ public class FolderUsage {
     public void setFolderID (long folderid) { this.folderid = folderid; }
     public void setPublicName (String publicname) { this.publicname = publicname; }
     public void setSharePass (String sharepass) { this.sharepass = sharepass; }
-    public void setCost (int cost) { costperuse = cost; }
+    public void setCost (float cost) { costperuse = cost; }
     //======================================================================
     public long projectID () { return projectid; }
     public long folderID () { return folderid; }
-    public int costPerUse () { return costperuse; }
+    public float costPerUse () { return costperuse; }
     public String publicName () {
         if (publicname == null) return "";
         return publicname;
